[ADD] weblate: Add support for weblate translations (#155)

+++ travis/test_server.py
@@ -587,7 +587,7 @@ def main(argv=None):
                 counted_errors += errors
                 all_errors.append(to_test)
                 print(fail_msg, "Found %d lines with errors" % errors)
-        if not instance_alive:
+        if not instance_alive and odoo_unittest:
             subprocess.call(["dropdb", database])
 
     print('Module test summary')
